[TH2-5132] behaviour.permittedToRemoveNamespace option * added logging and secret value check

+++ src/main/java/com/exactpro/th2/inframgr/repository/RepositoryWatcherService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2021 Exactpro (Exactpro Systems Limited)
+ * Copyright 2020-2023 Exactpro (Exactpro Systems Limited)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@ import com.exactpro.th2.inframgr.Config;
 import com.exactpro.th2.inframgr.SchemaEventRouter;
 import com.exactpro.th2.inframgr.docker.monitoring.DynamicResourceProcessor;
 import com.exactpro.th2.inframgr.k8s.K8sResourceCache;
+import com.exactpro.th2.inframgr.util.cfg.BehaviourCfg;
 import com.exactpro.th2.inframgr.util.cfg.GitCfg;
 import com.exactpro.th2.infrarepo.git.GitterContext;
 import io.fabric8.kubernetes.api.model.Namespace;
@@ -37,16 +38,20 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import static com.exactpro.th2.inframgr.SchemaController.SOURCE_BRANCH;
+
 @Component
 public class RepositoryWatcherService {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(RepositoryWatcherService.class);
+
     private final Map<String, String> commitHistory;
+
+    private final BehaviourCfg behaviour;
 
     private final GitCfg config;
 
     private final SchemaEventRouter eventRouter;
-
-    private final Logger logger;
 
     private final KubernetesClient kubeClient = new KubernetesClientBuilder().build();
 
@@ -57,20 +62,21 @@ public class RepositoryWatcherService {
     private static volatile boolean startupSynchronizationComplete;
 
     public RepositoryWatcherService() throws Exception {
-        var fullConfig = Config.getInstance();
+        Config fullConfig = Config.getInstance();
         commitHistory = new HashMap<>();
+        behaviour = fullConfig.getBehaviour();
         config = fullConfig.getGit();
         namespacePrefix = fullConfig.getKubernetes().getNamespacePrefix();
         eventRouter = SchemaEventRouter.getInstance();
-        logger = LoggerFactory.getLogger(RepositoryWatcherService.class);
     }
 
     @Scheduled(fixedDelayString = "${GIT_FETCH_INTERVAL:14000}")
     private void scheduledJob() {
         try {
-            logger.debug("fetching changes from git");
+            LOGGER.debug("fetching changes from git");
             GitterContext ctx = GitterContext.getContext(config);
             Map<String, String> commits = ctx.getAllBranchesCommits();
+            LOGGER.info("Fetched branches: {}, previous branch count: {}", commits, prevBranchCount);
             if (prevBranchCount > commits.size()) {
                 removeExtinctedNamespaces(commits.keySet());
             }
@@ -80,38 +86,38 @@ public class RepositoryWatcherService {
             }
             commits.forEach((branch, commitRef) -> {
 
-                if (!(branch.equals("master")
+                if (!(SOURCE_BRANCH.equals(branch)
                         || commitHistory.isEmpty()
                         || commitHistory.getOrDefault(branch, "").equals(commitRef))) {
-                    logger.info("New commit \"{}\" detected for branch \"{}\"", commitRef, branch);
+                    LOGGER.info("New commit \"{}\" detected for branch \"{}\"", commitRef, branch);
 
                     RepositoryUpdateEvent event = new RepositoryUpdateEvent(branch, commitRef);
                     boolean sent = eventRouter.addEventIfNotCached(branch, event);
                     if (!sent) {
-                        logger.info("Event is recently processed, ignoring");
+                        LOGGER.info("Event is recently processed, ignoring");
                     }
                 }
             });
 
             commitHistory.putAll(commits);
         } catch (Exception e) {
-            logger.error("Error fetching repository", e);
+            LOGGER.error("Error fetching repository", e);
         }
     }
 
     private void doInitialSynchronization(Map<String, String> commits) {
-        logger.info("Starting Kubernetes synchronization phase");
+        LOGGER.info("Starting Kubernetes synchronization phase");
         commits.forEach((branch, commitRef) -> {
-            if (!branch.equals("master")) {
+            if (!SOURCE_BRANCH.equals(branch)) {
                 RepositoryUpdateEvent event = new RepositoryUpdateEvent(branch, commitRef);
                 boolean sent = eventRouter.addEventIfNotCached(branch, event);
                 if (!sent) {
-                    logger.info("Event is recently processed, ignoring");
+                    LOGGER.info("Event is recently processed, ignoring");
                 }
             }
         });
         startupSynchronizationComplete = true;
-        logger.info("Kubernetes synchronization phase complete");
+        LOGGER.info("Kubernetes synchronization phase complete");
     }
 
     private void removeExtinctedNamespaces(Set<String> existingBranches) {
@@ -122,7 +128,7 @@ public class RepositoryWatcherService {
                 .map(item -> item.getMetadata().getName())
                 .filter(namespace -> namespace.startsWith(namespacePrefix)
                         && !existingBranches.contains(namespace.substring(namespacePrefix.length())))
-                .collect(Collectors.toList());
+                .toList();
 
         for (String extinctNamespace : extinctNamespaces) {
             String schemaName = extinctNamespace.substring(namespacePrefix.length());
@@ -132,9 +138,20 @@ public class RepositoryWatcherService {
             if (namespaceResource != null) {
                 String branchName = extinctNamespace.substring(namespacePrefix.length());
                 eventRouter.removeEventsForSchema(branchName);
-                logger.info("branch \"{}\" was removed from remote repository, deleting corresponding namespace \"{}\"",
-                        branchName, existingBranches);
-                namespaceResource.delete();
+                if (behaviour.isPermittedToRemoveNamespace()) {
+                    LOGGER.info(
+                            "branch \"{}\" was removed from remote repository, deleting corresponding namespace \"{}\"",
+                            branchName,
+                            existingBranches
+                    );
+                    namespaceResource.delete();
+                } else {
+                    LOGGER.warn(
+                            "branch \"{}\" was removed from remote repository, stopping namespace \"{}\" maintenance",
+                            branchName,
+                            existingBranches
+                    );
+                }
                 commitHistory.remove(branchName);
             }
         }
